Problem seems to be detecting collisions between two vertical lines: separate those out into their own tests
Results are a bit odd, we're detecting the initial collision only in the first and 3 collisions on the same point for some reason if you reverse the order. Lines are perfectly linear so we should really be detecting them as such but it looks like we're treating them as non-overlapping here.