The issue seems to be this filter
Removing it doesn't cause any other test failures, so I'm commenting it out for now, need to think about the implications more. Wondering if this was hiding the same bug in some cases.

The filter is weird as it stops the thing that moves the collisions on to the next element.